# FitnessArena-API/Models/post.cs
// <auto-generated> This file has been auto generated by EF Core Power Tools. </auto-generated>
#nullable disable
using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using System.ComponentModel.DataAnnotations.Schema;
using System.Text.Json.Serialization;
using Microsoft.EntityFrameworkCore;

namespace FitnessArena_API.Models
{
    [Table("post")]
    public partial class post
    {
        public post()
        {
            postsComments = new HashSet<postsComment>();
            postsLikes = new HashSet<postsLike>();
            postsShares = new HashSet<postsShare>();
        }

        [Key]
        public int postId { get; set; }
        [Required]
        [StringLength(500)]
        [Unicode(false)]
        public string content { get; set; }
        [Column(TypeName = "date")]
        public DateTime? postDate { get; set; }
        public int userid { get; set; }
        [StringLength(100)]
        [Unicode(false)]
        public string photoSrc { get; set; }
        [NotMapped]
        public int? numberOfLikes { get; set; }
        [NotMapped]
        public int? numberOfShares { get; set; }
        [NotMapped]
        public int? numberOfComments { get; set; }


        [ForeignKey("userid")]
        
        [InverseProperty("posts")]
        public virtual user user { get; set; }
        [JsonIgnore]
        [InverseProperty("post")]
        public virtual ICollection<postsComment> postsComments { get; set; }
        [JsonIgnore]
        [InverseProperty("post")]
        public virtual ICollection<postsLike> postsLikes { get; set; }
        [JsonIgnore]
        [InverseProperty("post")]
        public virtual ICollection<postsShare> postsShares { get; set; }
    }
}