# FitnessArena-API/Models/postsComment.cs
// <auto-generated> This file has been auto generated by EF Core Power Tools. </auto-generated>
#nullable disable
using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using System.ComponentModel.DataAnnotations.Schema;
using System.Text.Json.Serialization;
using Microsoft.EntityFrameworkCore;

namespace FitnessArena_API.Models
{
    [Table("postsComment")]
    public partial class postsComment
    {
        public postsComment()
        {
            commentLikes = new HashSet<commentLike>();
        }

        public int postId { get; set; }
        public int userId { get; set; }
        [JsonIgnore]
        [Key]
        public int commentId { get; set; }
        [StringLength(100)]
        [Unicode(false)]
        public string Content { get; set; }

        [ForeignKey("postId")]
        [JsonIgnore]
        [InverseProperty("postsComments")]
        public virtual post post { get; set; }
        [ForeignKey("userId")]
        [InverseProperty("postsComments")]
        [JsonIgnore]
        public virtual user user { get; set; }
        [JsonIgnore]
        [InverseProperty("comment")]
        public virtual ICollection<commentLike> commentLikes { get; set; }
    }
}