# FitnessArena-API/Models/user.cs
// <auto-generated> This file has been auto generated by EF Core Power Tools. </auto-generated>
#nullable disable
using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using System.ComponentModel.DataAnnotations.Schema;
using System.Text.Json.Serialization;
using Microsoft.EntityFrameworkCore;

namespace FitnessArena_API.Models
{
    [Index("email", Name = "UQ__users__AB6E616474F5F002", IsUnique = true)]
    public partial class user
    {
        public user()
        {
            Paymentcoaches = new HashSet<Payment>();
            Paymenttrainees = new HashSet<Payment>();
            bundles = new HashSet<bundle>();
            commentLikes = new HashSet<commentLike>();
            inbodies = new HashSet<inbody>();
            posts = new HashSet<post>();
            postsComments = new HashSet<postsComment>();
            postsLikes = new HashSet<postsLike>();
            postsShares = new HashSet<postsShare>();
            programs = new HashSet<program>();
            traineeBundles = new HashSet<traineeBundle>();
            traineeprograms = new HashSet<traineeprogram>();
        }

        [Key]
        public int userId { get; set; }
        [Required]
        [StringLength(10)]
        [Unicode(false)]
        public string type { get; set; }
        [Column(TypeName = "date")]
        public DateTime birthDate { get; set; }
        [Required]
        [StringLength(20)]
        [Unicode(false)]
        public string fname { get; set; }
        [Required]
        [StringLength(20)]
        [Unicode(false)]
        public string lname { get; set; }
        [StringLength(50)]
        [Unicode(false)]
        public string address { get; set; }
        [Column(TypeName = "date")]
        public DateTime? registerDate { get; set; }
        public int? age { get; set; }
        [Required]
        [StringLength(10)]
        [Unicode(false)]
        public string gender { get; set; }
        [Column(TypeName = "decimal(3, 2)")]
        public decimal? rating { get; set; }
        [StringLength(100)]
        [Unicode(false)]
        public string PhotoSrc { get; set; }
        [StringLength(20)]
        [Unicode(false)]
        public string CategoryName { get; set; }
        [StringLength(20)]
        [Unicode(false)]
        public string target { get; set; }
        public int? cardNumber { get; set; }
        [Required]
        [StringLength(50)]
        [Unicode(false)]
        public string email { get; set; }
        [Required]
        [StringLength(50)]
        [Unicode(false)]
        public string password { get; set; }

        [InverseProperty("coach")]
        [JsonIgnore]
        public virtual ICollection<Payment> Paymentcoaches { get; set; }
        [JsonIgnore]
        [InverseProperty("trainee")]
        public virtual ICollection<Payment> Paymenttrainees { get; set; }
        [JsonIgnore]
        [InverseProperty("coach")]
        public virtual ICollection<bundle> bundles { get; set; }
        [JsonIgnore]
        [InverseProperty("user")]
        public virtual ICollection<commentLike> commentLikes { get; set; }
        [JsonIgnore]
        [InverseProperty("trainee")]
        public virtual ICollection<inbody> inbodies { get; set; }
        [JsonIgnore]
        [InverseProperty("user")]
        public virtual ICollection<post> posts { get; set; }
        [JsonIgnore]
        [InverseProperty("user")]
        public virtual ICollection<postsComment> postsComments { get; set; }
        [JsonIgnore]
        [InverseProperty("user")]
        public virtual ICollection<postsLike> postsLikes { get; set; }
        [JsonIgnore]
        [InverseProperty("user")]
        public virtual ICollection<postsShare> postsShares { get; set; }
        [JsonIgnore]
        [InverseProperty("coach")]
        public virtual ICollection<program> programs { get; set; }
        [JsonIgnore]
        [InverseProperty("trainee")]
        public virtual ICollection<traineeBundle> traineeBundles { get; set; }
        [JsonIgnore]
        [InverseProperty("trainee")]
        public virtual ICollection<traineeprogram> traineeprograms { get; set; }
    }
}