# FitnessArena-API/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "FitnessArena-API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/bundles": {
      "get": {
        "tags": [
          "bundles"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/bundle"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "bundles"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/bundle"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/bundle"
            }
          }
        }
      }
    },
    "/api/bundles/{id}": {
      "get": {
        "tags": [
          "bundles"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/bundle"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "bundles"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/bundle"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "bundles"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/inbodies": {
      "get": {
        "tags": [
          "inbodies"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/inbody"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "inbodies"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/inbody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/inbody"
            }
          }
        }
      }
    },
    "/api/inbodies/trainee/{id}": {
      "get": {
        "tags": [
          "inbodies"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/inbody"
              }
            }
          }
        }
      }
    },
    "/api/inbodies/{id}": {
      "put": {
        "tags": [
          "inbodies"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/inbody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "inbodies"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/posts/{id}": {
      "get": {
        "tags": [
          "posts"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/post"
            }
          }
        }
      },
      "put": {
        "tags": [
          "posts"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/post"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "posts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/postsComments": {
      "get": {
        "tags": [
          "postsComments"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/postsComment"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "postsComments"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/postsComment"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/postsComment"
            }
          }
        }
      }
    },
    "/api/postsComments/{id}": {
      "get": {
        "tags": [
          "postsComments"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "postsComments"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/postsComment"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "postsComments"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "bundle": {
      "required": [
        "name"
      ],
      "type": "object",
      "properties": {
        "bundleId": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "maxLength": 10,
          "minLength": 0,
          "type": "string"
        },
        "fees": {
          "format": "int32",
          "type": "integer"
        },
        "coachId": {
          "format": "int32",
          "type": "integer"
        },
        "durationDays": {
          "format": "int32",
          "type": "integer"
        },
        "description": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "coach": {
          "$ref": "#/definitions/user"
        },
        "traineeBundles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/traineeBundle"
          }
        }
      }
    },
...
    "inbody": {
      "type": "object",
      "properties": {
        "inbodyId": {
          "format": "int32",
          "type": "integer"
        },
        "date": {
          "format": "date-time",
          "type": "string"
        },
        "muscle": {
          "format": "double",
          "type": "number"
        },
        "minerals": {
          "format": "double",
          "type": "number"
        },
        "proteins": {
          "format": "double",
          "type": "number"
        },
        "water": {
          "format": "double",
          "type": "number"
        },
        "weight": {
          "format": "double",
          "type": "number"
        },
        "height": {
          "format": "int32",
          "type": "integer"
        },
        "traineeId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "post": {
      "required": [
        "content"
      ],
      "type": "object",
      "properties": {
        "postId": {
          "format": "int32",
          "type": "integer"
        },
        "content": {
          "maxLength": 500,
          "minLength": 0,
          "type": "string"
        },
        "postDate": {
          "format": "date-time",
          "type": "string"
        },
        "userid": {
          "format": "int32",
          "type": "integer"
        },
        "photoSrc": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "numberOfLikes": {
          "format": "int32",
          "type": "integer"
        },
        "numberOfShares": {
          "format": "int32",
          "type": "integer"
        },
        "numberOfComments": {
          "format": "int32",
          "type": "integer"
        },
        "user": {
          "$ref": "#/definitions/user"
        }
      }
    },
    "postsComment": {
      "type": "object",
      "properties": {
        "postId": {
          "format": "int32",
          "type": "integer"
        },
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "content": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        }
      }
    },
...
    "traineeBundle": {
      "type": "object",
      "properties": {
        "bundleId": {
          "format": "int32",
          "type": "integer"
        },
        "traineeId": {
          "format": "int32",
          "type": "integer"
        },
        "subscriptionDate": {
          "format": "date-time",
          "type": "string"
        },
        "expiryDate": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
...
    "user": {
      "required": [
        "email",
        "fname",
        "gender",
        "lname",
        "password",
        "type"
      ],
      "type": "object",
      "properties": {
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "type": {
          "maxLength": 10,
          "minLength": 0,
          "type": "string"
        },
        "birthDate": {
          "format": "date-time",
          "type": "string"
        },
        "fname": {
          "maxLength": 20,
          "minLength": 0,
          "type": "string"
        },
        "lname": {
          "maxLength": 20,
          "minLength": 0,
          "type": "string"
        },
        "address": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "registerDate": {
          "format": "date-time",
          "type": "string"
        },
        "age": {
          "format": "int32",
          "type": "integer"
        },
        "gender": {
          "maxLength": 10,
          "minLength": 0,
          "type": "string"
        },
        "rating": {
          "format": "double",
          "type": "number"
        },
        "photoSrc": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "categoryName": {
          "maxLength": 20,
          "minLength": 0,
          "type": "string"
        },
        "target": {
          "maxLength": 20,
          "minLength": 0,
          "type": "string"
        },
        "cardNumber": {
          "format": "int32",
          "type": "integer"
        },
        "email": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "password": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        }
      }
    },
...
  }
}